requirements_databricks.txt
cfgv==3.4.0
distlib==0.4.0
filelock==3.19.1
identify==2.6.14
importlib_metadata==8.7.0
nodeenv==1.9.1
platformdirs==4.4.0
pre_commit==4.3.0
python-dotenv==1.1.1
PyYAML==6.0.2
typing_extensions==4.15.0
virtualenv==20.34.0
zipp==3.23.0
